Move appData to native/
