+++ Tools/WinML-Dashboard/src/native/appData.tsx
