+++ Tools/WinML-Dashboard/src/view/convert/View.tsx
@@ -10,10 +10,10 @@ import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import Collapsible from '../../components/Collapsible';
 import { setFile, setSaveFileName } from '../../datastore/actionCreators';
 import IState from '../../datastore/state';
+import { packagedFile } from '../../native/appData';
 import { fileFromPath, showNativeOpenDialog, showNativeSaveDialog } from '../../native/dialog';
 import { downloadPip, downloadPython, getLocalPython, getPythonBinaries, installVenv, pip, python } from '../../native/python';
 import { isWeb } from '../../native/util';
-import { packagedFile } from '../../persistence/appData';
 
 import './View.css';
